fix(theme): to always merge theme options with default options
