+++ packages/gatsby-theme-docs/default-options.js
@@ -1,0 +1,14 @@
+const { customProperties } = require('@commercetools-uikit/design-system');
+
+const defaultOptions = {
+  websiteKey: '',
+  websitePrimaryColor: customProperties.colorPrimary,
+  beta: false,
+  gaTrackingId: undefined,
+  excludeFromSearchIndex: true,
+  createNodeSlug: undefined,
+  additionalPrismLanguages: [],
+  overrideDefaultConfigurationData: [],
+};
+
+module.exports = defaultOptions;
